[REFACTOR/#12] 패키지 경로 수정
