+++ feature/src/main/java/com/terning/feature/home/home/component/HomeFilteringScreen.kt
@@ -1,4 +1,4 @@
-package com.terning.feature.home.component
+package com.terning.feature.home.home.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
