+++ feature/src/main/java/com/terning/feature/home/home/component/HomeTodayIntern.kt
@@ -1,4 +1,4 @@
-package com.terning.feature.home.component
+package com.terning.feature.home.home.component
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.PaddingValues
