+++ feature/src/main/java/com/terning/feature/home/home/component/HomeTodayInternItem.kt
@@ -1,4 +1,4 @@
-package com.terning.feature.home.component
+package com.terning.feature.home.home.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
