+++ feature/src/main/java/com/terning/feature/home/home/navigation/HometNavigation.kt
@@ -1,11 +1,11 @@
-package com.terning.feature.home.navigation
+package com.terning.feature.home.home.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.terning.core.navigation.MainTabRoute
-import com.terning.feature.home.HomeRoute
+import com.terning.feature.home.home.HomeRoute
 import kotlinx.serialization.Serializable
 
 fun NavController.navigateHome(navOptions: NavOptions? = null) {
