+++ feature/src/main/java/com/terning/feature/main/MainNavigator.kt
@@ -10,9 +10,9 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navOptions
 import com.terning.feature.calendar.navigation.navigateCalendar
-import com.terning.feature.home.navigation.navigateHome
+import com.terning.feature.home.home.navigation.navigateHome
 import com.terning.feature.mypage.navigation.navigateMyPage
-import com.terning.feature.search.search.navigation.Search
+import com.terning.feature.onboarding.signin.navigation.SignIn
 import com.terning.feature.search.search.navigation.navigateSearch
 
 class MainNavigator(
@@ -22,7 +22,7 @@ class MainNavigator(
         @Composable get() = navController
             .currentBackStackEntryAsState().value?.destination
 
-    val startDestination = Search
+    val startDestination = SignIn
 
     val currentTab: MainTab?
         @Composable get() = MainTab.find { tab ->
